+++ test/simply_stored_couch_test.rb
@@ -1295,7 +1295,7 @@ class CouchTest < Test::Unit::TestCase
       context "when saving the attachment" do
         should "fetch the collection" do
           @log_item.log_data = "Yay! It logged!"
-          RightAws::S3.expects(:new).with('abcdef', 'secret!', :multi_thread => true, :ca_file => nil).returns(@s3)
+          RightAws::S3.expects(:new).with('abcdef', 'secret!', :multi_thread => true, :ca_file => nil, :logger => nil).returns(@s3)
           @log_item.save
         end
       
@@ -1348,6 +1348,16 @@ class CouchTest < Test::Unit::TestCase
           assert_raise(ArgumentError) do
             @log_item.save
           end
+        end
+        
+        should "pass the logger object down to RightAws" do
+          logger = mock()
+          @log_item.log_data = "Yay! log me"
+          CouchLogItem._s3_options[:log_data][:bucket] = 'mybucket'
+          CouchLogItem._s3_options[:log_data][:logger] = logger
+          
+          RightAws::S3.expects(:new).with(anything, anything, {:logger => logger, :ca_file => nil, :multi_thread => true}).returns(@s3)
+          @log_item.save
         end
       
         should "not upload the attachment when it hasn't been changed" do
@@ -1430,6 +1440,33 @@ class CouchTest < Test::Unit::TestCase
           @bucket.expects(:put).with(anything, '["one log entry","and another one"]', {}, anything)
           @log_item.save
         end
+        
+        context "when noting the size of the attachment" do
+          should "store on upload" do
+            @log_item.log_data = 'abc'
+            @bucket.expects(:put)
+            assert @log_item.save
+            assert_equal 3, @log_item.log_data_size
+          end
+        
+          should "update the size if the attachment gets updated" do
+            @log_item.log_data = 'abc'
+            @bucket.stubs(:put)
+            assert @log_item.save
+            assert_equal 3, @log_item.log_data_size
+          
+            @log_item.log_data = 'example'
+            assert @log_item.save
+            assert_equal 7, @log_item.log_data_size
+          end
+          
+          should "store the size of json attachments" do
+            @log_item.log_data = ['abc']
+            @bucket.stubs(:put)
+            assert @log_item.save
+            assert_equal ['abc'].to_json.size, @log_item.log_data_size
+          end
+        end
       end
     
       context "when fetching the data" do
@@ -1438,7 +1475,7 @@ class CouchTest < Test::Unit::TestCase
           CouchLogItem._s3_options[:log_data][:location] = :eu
           CouchLogItem._s3_options[:log_data][:ca_file] = '/etc/ssl/ca.crt'
           
-          RightAws::S3.expects(:new).with('abcdef', 'secret!', :multi_thread => true, :ca_file => '/etc/ssl/ca.crt').returns(@s3)
+          RightAws::S3.expects(:new).with('abcdef', 'secret!', :multi_thread => true, :ca_file => '/etc/ssl/ca.crt', :logger => nil).returns(@s3)
           
           @log_item.log_data
         end
@@ -1901,7 +1938,6 @@ class CouchTest < Test::Unit::TestCase
       
       context "when counting" do
         setup do
-          recreate_db
           @hemorrhoid = Hemorrhoid.create(:nickname => 'Claas')
           assert @hemorrhoid.destroy
           assert @hemorrhoid.reload.deleted?
@@ -1934,6 +1970,97 @@ class CouchTest < Test::Unit::TestCase
         end  
       end
 
+    end
+    
+    context "when handling conflicts" do
+      setup do
+        @original = User.create(:name => 'Mickey Mouse', :title => "Dr.", :homepage => 'www.gmx.de')
+        @copy = User.find(@original.id)
+        User.auto_conflict_resolution_on_save = true
+      end
+      
+      should "be able to save without modifications" do
+        assert @copy.save
+      end
+      
+      should "be able to save when modification happen on different attributes" do
+        @original.name = "Pluto"
+        assert @original.save
+        
+        @copy.title = 'Prof.'
+        assert_nothing_raised do
+          assert @copy.save
+        end
+        
+        assert_equal "Pluto", @copy.reload.name
+        assert_equal "Prof.", @copy.reload.title
+        assert_equal "www.gmx.de", @copy.reload.homepage
+      end
+      
+      should "be able to save when modification happen on different, multiple attributes - remote" do
+        @original.name = "Pluto"
+        @original.homepage = 'www.google.com'
+        assert @original.save
+        
+        @copy.title = 'Prof.'
+        assert_nothing_raised do
+          assert @copy.save
+        end
+        
+        assert_equal "Pluto", @copy.reload.name
+        assert_equal "Prof.", @copy.reload.title
+        assert_equal "www.google.com", @copy.reload.homepage
+      end
+      
+      should "be able to save when modification happen on different, multiple attributes locally" do
+        @original.name = "Pluto"
+        assert @original.save
+        
+        @copy.title = 'Prof.'
+        @copy.homepage = 'www.google.com'
+        assert_nothing_raised do
+          assert @copy.save
+        end
+        
+        assert_equal "Pluto", @copy.reload.name
+        assert_equal "Prof.", @copy.reload.title
+        assert_equal "www.google.com", @copy.reload.homepage
+      end
+      
+      should "re-raise the conflict if there is no merge possible" do
+        @original.name = "Pluto"
+        assert @original.save
+        
+        @copy.name = 'Prof.'
+        assert_raise(RestClient::Conflict) do
+          assert @copy.save
+        end
+        
+        assert_equal "Prof.", @copy.name
+        assert_equal "Pluto", @copy.reload.name
+      end
+      
+      should "re-raise the conflict if retried several times" do
+        exception = RestClient::Conflict.new
+        CouchPotato.database.expects(:save_document).raises(exception).times(3)
+        
+        @copy.name = 'Prof.'
+        assert_raise(RestClient::Conflict) do
+          assert @copy.save
+        end
+      end
+      
+      should "not try to merge and re-save if auto_conflict_resolution_on_save is disabled" do
+        User.auto_conflict_resolution_on_save = false
+        exception = RestClient::Conflict.new
+        CouchPotato.database.expects(:save_document).raises(exception).times(1)
+        
+        @copy.name = 'Prof.'
+        assert_raise(RestClient::Conflict) do
+          assert @copy.save
+        end
+      end
+      
     end
     
   end