+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 Changelog
 =============
 
+0.3.0
+=============
+
+- SimplyStored now automatically retries conflicted save operations if it is possible to resolve the conflict.
+  Solving the conflict means that if updated were done one different attributes the local object will 
+  refresh those attributes and try to save again. This will be tried two times by default. Afterwards the conflict
+  exception will be re-raised.
+  
+  This feature can be controlled on the class level like this: User.auto_conflict_resolution_on_save = true | false
+
+0.2.5
+=============
+
+- Allow to pass a custom logger to S3-attachments / RightAws
+
 - Support :order option to has_many associations and classes, e.g:
 
     @user.posts(:order => :desc)